Uses Rx BehaviorSubject to exchange radio filtering events
- Fixes Rollup config removing un-necessary external references
- Removes unnecessary maputils service

Creates a BehaviorSubject and uses that to propagate the values
of the strongest and latest quakes as aggregated by scan. The
change event is filtered by mag or time to then propagate to
the subject the corresponding strongest or latest quake object.
The map view is then set in the map.component to the quake lat
and lng.